requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.10.7
Flask==2.2.3
fonttools==4.39.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sympy==1.11.1
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
gunicorn
